refactor(gh-pages): landing components restructuring
