+++ pages/src/landing/benefits/benefits.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 .benefit {
   position: relative;
