+++ pages/src/landing/browser-support/browser-support.less
@@ -1,3 +1,5 @@
+@import (reference) "../../common/variables.less";
+
 .browser-support {
   position: relative;
   flex-direction: column;
