+++ pages/src/landing/landing.ts
@@ -1,0 +1,1 @@
+export * from './marquee/marquee';
