+++ pages/src/landing/marquee/marquee.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 .marquee {
   position: relative;
