+++ pages/src/landing/marquee/marquee.ts
@@ -1,9 +1,9 @@
-import {ESLBaseElement, attr} from '../../../src/modules/esl-base-element/core';
-import {bind} from '../../../src/modules/esl-utils/decorators/bind';
-import {ready} from '../../../src/modules/esl-utils/decorators/ready';
-import {range} from '../../../src/modules/esl-utils/misc/array';
-import {memoize} from '../../../src/modules/esl-utils/decorators/memoize';
-import {isIE} from '../../../src/modules/esl-utils/environment/device-detector';
+import {ESLBaseElement, attr} from '../../../../src/modules/esl-base-element/core';
+import {bind} from '../../../../src/modules/esl-utils/decorators/bind';
+import {ready} from '../../../../src/modules/esl-utils/decorators/ready';
+import {range} from '../../../../src/modules/esl-utils/misc/array';
+import {memoize} from '../../../../src/modules/esl-utils/decorators/memoize';
+import {isIE} from '../../../../src/modules/esl-utils/environment/device-detector';
 
 export class ESLDemoMarquee extends ESLBaseElement {
   static is = 'esl-d-marquee';
