+++ pages/src/landing/overview/overview-item.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 .overview {
   &-item {
