+++ pages/src/landing/overview/overview.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 .overview {
   position: relative;
