+++ pages/src/localdev.less
@@ -21,11 +21,9 @@
 
 @import "./navigation/navigation.less";
 @import "./related/related.less";
-@import "./marquee/marquee.less";
+@import "./landing/landing.less";
+
 @import "./collection-grid/collection-grid.less";
-@import "./benefits/benefits.less";
-@import "./browser-support/browser-support.less";
-@import "./overview/overview.less";
 @import "./esl-media-demo/test-media.less";
 
 @import "../../src/modules/all.less";
